+++ CHANGELOG.md
@@ -5,6 +5,16 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2023-08-18
+
+### Fixed
+
+- Default cookie option for `SameSite` (strict) wasn't set on the client.
+
+### Added
+
+- `flashCookieOptions` added to `getFlash` options.
+
 ## [2.1.3] - 2023-08-09
 
 ### Fixed